android_media_MediaMuxer_setLocation: amended signature.
Change parameter `nativeObject` from type `jint` to `jlong` to match its JNI
signature.

Bug: 15337287
Bug: 12890910
(cherry picked from commit 3cb78498d3f664f541ba7d28f4543cf8c12733f1)

+++ media/jni/android_media_MediaMuxer.cpp
@@ -164,7 +164,7 @@ static void android_media_MediaMuxer_setOrientationHint(
 }
 
 static void android_media_MediaMuxer_setLocation(
-        JNIEnv *env, jclass clazz, jint nativeObject, jint latitude, jint longitude) {
+        JNIEnv *env, jclass clazz, jlong nativeObject, jint latitude, jint longitude) {
     MediaMuxer* muxer = reinterpret_cast<MediaMuxer *>(nativeObject);
 
     status_t res = muxer->setLocation(latitude, longitude);
